Add .gitignore and add ability to manually generate weights and biases in a layer

+++ src/main/java/org/example/Bruh.java
@@ -2,7 +2,7 @@ package org.example;
 
 public interface Bruh {
 
-	float[] calculateOutputs();
+	double[] calculateOutputs();
 	void showContents();
 
 }
